refactor: PR#118のコメントに対応したリファクタリング
【Domain層】
- RouteEntityをMissionEntityにリネーム
- LocationEntityを削除し、Coordinate値オブジェクトを使用
- ImageCoordinateでCoordinate値オブジェクトを使用
- Radius値オブジェクトを追加（500m-10000mの制約）
- 値オブジェクトにバリデーションを追加（Coordinate: 緯度-90〜90、経度-180〜180）
- null許容型を削除し、必須フィールドをrequiredに変更
- mission_model.dartを削除し、entity/とvalue_object/に分離

【Application層】
- インターフェースにIプレフィックスを追加（ILocationService、IMissionRepository）
- mission_repository_interface.dartをmission_repository.dartにリネーム
- location_service_interface.dartをlocation_service.dartにリネーム
- GetCurrentLocationUseCaseを削除（GetMissionUseCaseがILocationServiceを直接使用）
- GetMissionUseCaseを更新（Radius値オブジェクトを受け取るように変更）

【Data層】
- LocationServiceを更新（Coordinateを返すように変更）
- MissionRepositoryを更新（Radius値オブジェクトとCoordinate値オブジェクトを使用）

【Presentation層】
- mission_storeをRiverpod familyパターンに変更（buildメソッドでRadiusを受け取る）
- mission_pageを更新（Radius値オブジェクトを保持、familyパターンに対応）
- setup_pageを更新（Radius値オブジェクトを使用）

【DI層】
- mission_provider.dartを新規作成（DIと状態管理を分離）
- di/ディレクトリを新規作成し、mission_provider.dartを配置（他の層と並列構造に統一）

【Core層】
- app_router.dartをrouter.dartに移動（core/router.dart）
- router.dartを更新（domain層に依存しないようにint型を使用）

+++ frontend/lib/core/router.dart
@@ -18,7 +18,7 @@ final GoRouter appRouter = GoRouter(
     GoRoute(
       path: '/mission/:radius',
       builder: (context, state) {
-        final radius = double.parse(state.pathParameters['radius']!);
+        final radius = int.parse(state.pathParameters['radius']!);
         return MissionPage(radius: radius);
       },
     ),
